fix(condition-tree): allow empty list with In operator
